fix: Make the Authorization Key's Size Dynamic
The Bearer token grows proportionally to the amount of information it
holds, making it difficult to estimate its size accurately. Before this
patch, while computing the headers to be sent to the server, the
implementation assumed a fixed value that become insufficient for holding
Azure tokens, leading to an API error.

This patch dynamically calculates the size of the token based on the token
environment variable, accommodating current, past, and future sizes.

Note: (a) The implementation does not free all resources after the curl's
slist usage. (b) http client could be an abstraction; Those are
subject of different patches.